pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "conclave-agents"
version = "0.1.0"
description = "A bureau of Claude agents that deliberates, decides, and delivers."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Conclave Contributors" }]
keywords = ["anthropic", "claude", "agents", "multi-agent", "llm", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.30.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "respx>=0.20",
]
dashboard = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "watchfiles>=0.21",
]

[project.scripts]
conclave = "conclave.cli:cli"

[project.urls]
Homepage = "https://github.com/your-handle/conclave"
Documentation = "https://github.com/your-handle/conclave#readme"
Issues = "https://github.com/your-handle/conclave/issues"

[tool.hatch.build.targets.wheel]
packages = ["conclave"]
